scx_layered: s/monitor.rs/stats.rs/ and make stats refresh code struct ops
